refactor: extract memory.js from main.js (BAT-198)
Move ~260 lines of soul/memory/heartbeat code into dedicated memory.js module:
- DEFAULT_SOUL template, loadSoul(), loadBootstrap(), loadIdentity(), loadUser()
- loadMemory(), saveMemory(), getDailyMemoryPath(), loadDailyMemory(), appendDailyMemory()
- searchMemory() with SQL.js ranking + file-based fallback
- updateHeartbeat() with 5-minute interval

Database reference injected via setDb() getter pattern (same as setRedactFn).
Depends on config.js. No logic changes.
